pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "biometric_mlops"
version = "1.0.0"
description = "Multimodal biometric recognition MLOps pipeline"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Biometric MLOps Team"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "numpy>=1.24.0",
    "Pillow>=9.0.0",
    "pyarrow>=12.0.0",
    "pandas>=2.0.0",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]
ray = [
    "ray>=2.0.0",
]
mlflow = [
    "mlflow>=2.10.0",
]
monitoring = [
    "evidently>=0.4.0",
    "scipy>=1.10.0",
]
kaggle = [
    "kaggle>=1.5.0",
]
databricks = [
    "databricks-sdk>=0.20.0",
    "mlflow>=2.10.0",
]
all = [
    "biometric_mlops[dev,ray,mlflow,databricks,monitoring,kaggle]",
]

[project.scripts]
biometric-train = "src.train:main"
biometric-inference = "src.inference:main"
biometric-evaluate = "src.evaluate:main"
biometric-retrain = "src.retrain:main"
biometric-monitor = "src.monitoring:main"
biometric-download = "src.data_download:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
exclude = ["tests*"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.venv
    | build
    | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip = [".venv", "build", "dist"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
exclude = ["tests", "build", "dist"]

[tool.coverage.run]
source = ["src"]
omit = ["tests/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
